- mocked window.alert to pre-empt test error - added some missing local vue items - changed some mounts to shallowMounts - reset some store data to prevent conflicts between describe sections - updated some comments
